pyproject.toml
[tool.poetry]
name = "irc_gpt_bot"
version = "0.1.0"
description = ""
authors = ["Luc Léonard <github@luc-leonard.fr>"]

[tool.poetry.dependencies]
python = "3.9.0"
transformers = "^4.3.2"
torch = "^1.7.1"
irc = "^19.0.1"
praw = "^7.1.4"
psaw = "^0.1.0"
progressbar2 = "^3.53.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
